Selected Date Edit Function
Form refreshes when a different data is selected and passed the date.  Still need to configure the journal to load the values from that day.  There will be a small kink because if you have removed the value from the config file, it won't be loaded to edit, but I think I'll leave that and only load the fields that are current since that's what the ML functions will be using